Track 3 change in float 32 to float 16 for GPU

+++ track3_frame_generation/dreampose_like/infer_controlnet_img2img.py
@@ -30,7 +30,8 @@ def build_pipelines(device, sd_model, controlnet_model, dtype):
     pipe = StableDiffusionControlNetImg2ImgPipeline.from_pretrained(
     sd_model,
     controlnet=controlnet,
-    torch_dtype=torch.float32
+    # torch_dtype=torch.float32
+    torch_dtype=torch.float16
 )
 
     pipe.to(device)
